Make zombies slow down & respawn on low FPS
This update makes zombies think slower when the game is lagging, and if it continues to lag, they will respawn, and they will keep respawning until the lag stops. The reason these changes were added is because zombies are the #1 source of lag. This makes it near-impossible for zombies to single-handedly lag out the game.