Make comment less hostile
This comment was extremely insensitive and unnecessary. It was created out of frustration as a server hoster who was confused for several months as to why some people were getting a powerful weapon that no one else had. When I discovered the reasoning, I very quickly wrote the comment, but rather making it informative, it cursed instead. I do recognize that without Zet0r, we would not have such an amazing game to play, so making hostile comments like this in the very mode that we forked is admittedly stupid.